requirements.txt
# BiRefNet Background Removal Worker

# RunPod SDK
runpod>=1.6.0

# rembg with GPU support (includes BiRefNet)
rembg[gpu]>=2.0.50

# ONNX Runtime GPU
onnxruntime-gpu>=1.16.0

# Image handling
Pillow>=10.0.0
requests>=2.28.0

# Required by rembg
numpy>=1.24.0
opencv-python-headless>=4.8.0
